+++ src/object/vector/rep.rs
@@ -1,5 +1,5 @@
 use std::cell::RefCell;
-use std::fmt::{Debug, Display};
+use std::fmt::Debug;
 use std::rc::Rc;
 
 use super::coercion::{AtomicMode, CoercibleInto, CommonCmp, CommonNum, MinimallyNumeric};
@@ -294,7 +294,8 @@ impl<T: AtomicMode + Clone + Default> Rep<T> {
             .collect()
     }
 
-    fn get_inner(&self, index: usize) -> Option<T> {
+    // FIXME: private
+    pub fn get_inner(&self, index: usize) -> Option<T> {
         match self {
             Rep::Subset(v, subsets) => {
                 let vc = v.clone();
@@ -386,71 +387,6 @@ where
         match Self::from(value.0) {
             Rep::Subset(v, _) => Rep::Subset(v, value.1),
         }
-    }
-}
-
-impl<T> Display for Rep<T>
-where
-    T: AtomicMode + Debug + Default + Clone,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let n = self.len();
-        if n == 0 {
-            if self.is_double() {
-                return write!(f, "double(0)");
-            }
-            if self.is_integer() {
-                return write!(f, "integer(0)");
-            }
-            if self.is_logical() {
-                return write!(f, "logical(0)");
-            }
-            if self.is_character() {
-                return write!(f, "character(0)");
-            }
-        }
-
-        let nlen = format!("{}", n).len();
-        // TODO: iteratively calculate when we hit max print so our
-        // max_len isn't inflated by a value that is omitted
-
-        let xc = self.inner().clone();
-        let xb = xc.borrow();
-
-        let x_strs = xb.iter().map(|xi| format!("{:?}", xi));
-        let max_len = x_strs
-            .clone()
-            .fold(0, |max_len, xi| std::cmp::max(max_len, xi.len()));
-
-        let mut col = 0;
-        let gutterlen = 2 + nlen + 1;
-
-        // hard coded max print & console width
-        let maxprint = 20 * ((80 - gutterlen) / max_len);
-
-        x_strs
-            .take(maxprint)
-            .enumerate()
-            .try_for_each(|(i, x_str)| {
-                if i == 0 {
-                    col = gutterlen + max_len;
-                    write!(f, "{:>3$}[{}] {:>4$}", "", i + 1, x_str, nlen - 1, max_len)
-                } else if col + 1 + max_len > 80 {
-                    col = gutterlen + max_len;
-                    let i_str = format!("{}", i + 1);
-                    let gutter = nlen - i_str.len();
-                    write!(f, "\n{:>3$}[{}] {:>4$}", "", i_str, x_str, gutter, max_len)
-                } else {
-                    col += 1 + max_len;
-                    write!(f, " {:>1$}", x_str, max_len)
-                }
-            })?;
-
-        if n > maxprint {
-            write!(f, "\n[ omitting {} entries ]", n - maxprint)?;
-        }
-
-        Ok(())
     }
 }